Fix inconsistent naming. Rename recaptcha to requester
